+++ HALMETPreferences.h
@@ -4,6 +4,7 @@
 #include <Preferences.h>
 #include "DS18B20Processor.h"
 #include "VDOProcessor.h"
+#include "WaterProcessor.h"
 
 // === H A L M E T P R E F E R E N C E S  C L A S S  ( S K E L E T O N ) ===
 //
@@ -11,7 +12,9 @@
 
 class HALMETPreferences {
 public:
-    explicit HALMETPreferences(DS18B20Processor &ds18b20_proc, VDOProcessor &vdo_proc);
+    explicit HALMETPreferences(DS18B20Processor &ds18b20_proc,
+                               VDOProcessor     &vdo_proc,
+                               WaterProcessor   &water_proc);
 
     void load();  // placeholder
     void save();  // placeholder
@@ -19,5 +22,6 @@ public:
 private:
     DS18B20Processor &_ds18b20_proc;
     VDOProcessor     &_vdo_proc;
+    WaterProcessor   &_water_proc;
     Preferences       _prefs;
 };